Update to digikam 7.4.0 release.
Changes:
* Update baseimage-gui to use Debian 11; required by Digikam.
* Mitigate CVE-2021-44228 via Debian 11.
* Custom baseimage-gui build until a public Debian 11 release occurs.

The 7.4.0 release requires Debian 11 as well as mitigating CVE-2021-44228.

Debian 11 baseimage has not been released stable yet; therefore a custom build
has been created. As such, 7.4.0 will not be in the 'stable' release until
it is publically released and the build is reproducible without patches.

Build reproduction instructions are located in Dockerfile.